+++ src/components/branch-pages/WhyChooseDataAnalytics.jsx
@@ -1,0 +1,21 @@
+import React from "react";
+import { Typography } from "@mui/material";
+
+export default function WhyChooseDataAnalytics() {
+  return (
+    <Typography
+      textAlign='justify'
+      variant='body1'
+      fontWeight='bold'
+      sx={{ padding: "0px 3%" }}
+    >
+      At Pranika Technologies and Consulting Pvt. Ltd., we specialize in
+      delivering tailored data analytics solutions that empower businesses to
+      unlock the full potential of their data. Partner with us to transform your
+      data into a powerful asset that fuels your business growth. Explore our
+      success stories, discover how we’ve helped other businesses, and see what
+      we can do for you. Contact us today for a consultation and take the first
+      step towards smarter, data-driven decision-making.
+    </Typography>
+  );
+}
